requirements.txt
pandas
numpy
scikit-learn
tensorflow
matplotlib
yfinance
python-dotenv